There's nothing in the past. It's the future that has things to offer.
