+++ fractional-NFT/project/contracts/token-distributor-contract/tests/functions/core/buyback.rs
@@ -1,27 +1,28 @@
 use crate::utils::{
-    asset_abi_calls::mint_and_send_to_address,
-    nft_abi_calls::{approve, mint},
-    test_helpers::{defaults, setup, wallet_balance},
-    token_distributor_abi_calls::{buyback, create, purchase, sell},
+    interface::core::{
+        asset::mint_and_send_to_address,
+        nft::{approve, mint},
+        token_distributor::{buyback, create, purchase},
+    },
+    setup::{defaults, setup},
 };
-use fuels::{
-    prelude::{Address, Bech32ContractId, CallParameters, TxParameters},
-    signers::Signer,
-    tx::AssetId,
-    types::Identity,
-};
+use fuels::{prelude::Address, types::Identity};
 
 mod success {
 
     use super::*;
+    use crate::utils::{
+        interface::info::token_distributor::token_distribution, setup::DistributionState,
+    };
+    use fuels::{prelude::Bech32ContractId, tx::AssetId};
 
     #[tokio::test]
-    async fn sells() {
+    async fn starts_buyback() {
         let (
             deployer,
             owner1,
             owner2,
-            _token_distributor_contract,
+            token_distributor_contract,
             fractional_nft_contract,
             nft_contract,
             asset_contract,
@@ -67,6 +68,24 @@ mod success {
             token_price,
         )
         .await;
+
+        let token_distribution_struct =
+            token_distribution(&owner1.token_distributor, fractional_nft_contract).await;
+        assert_eq!(
+            provider
+                .get_contract_asset_balance(
+                    &Bech32ContractId::from(token_distributor_contract),
+                    AssetId::from(*asset_contract)
+                )
+                .await
+                .unwrap(),
+            purchase_amount * token_price
+        );
+        assert!(matches!(
+            token_distribution_struct.clone().unwrap().state,
+            DistributionState::Distributed()
+        ));
+
         buyback(
             purchase_amount * token_price,
             &owner1.token_distributor,
@@ -76,59 +95,31 @@ mod success {
         )
         .await;
 
-        assert_eq!(
-            wallet_balance(asset_contract, &owner2.wallet).await,
-            asset_supply - (purchase_amount * token_price)
-        );
-        assert_eq!(
-            wallet_balance(fractional_nft_contract, &owner2.wallet).await,
-            purchase_amount
-        );
+        let token_distribution_struct =
+            token_distribution(&owner1.token_distributor, fractional_nft_contract).await;
         assert_eq!(
             provider
                 .get_contract_asset_balance(
-                    &Bech32ContractId::from(fractional_nft_contract),
-                    AssetId::from(*fractional_nft_contract)
+                    &Bech32ContractId::from(token_distributor_contract),
+                    AssetId::from(*asset_contract)
                 )
                 .await
                 .unwrap(),
-            0
+            (purchase_amount * token_price) * 2
         );
-
-        sell(
-            purchase_amount,
-            &owner2.token_distributor,
-            fractional_nft_contract,
-        )
-        .await;
-
-        assert_eq!(
-            wallet_balance(asset_contract, &owner2.wallet).await,
-            asset_supply
-        );
-        assert_eq!(
-            wallet_balance(fractional_nft_contract, &owner2.wallet).await,
-            0
-        );
-        assert_eq!(
-            provider
-                .get_contract_asset_balance(
-                    &Bech32ContractId::from(fractional_nft_contract),
-                    AssetId::from(*fractional_nft_contract)
-                )
-                .await
-                .unwrap(),
-            purchase_amount
-        );
+        assert!(matches!(
+            token_distribution_struct.clone().unwrap().state,
+            DistributionState::Buyback()
+        ));
     }
 
     #[tokio::test]
-    async fn sells_some() {
+    async fn starts_return_at_different_price() {
         let (
-            _deployer,
+            deployer,
             owner1,
             owner2,
-            _token_distributor_contract,
+            token_distributor_contract,
             fractional_nft_contract,
             nft_contract,
             asset_contract,
@@ -138,6 +129,7 @@ mod success {
 
         let owner_identity = Identity::Address(owner1.wallet.address().into());
         let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
+        let provider = deployer.wallet.get_provider().unwrap();
 
         mint(1, &owner1.nft, owner_identity.clone()).await;
         approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
@@ -173,39 +165,49 @@ mod success {
             token_price,
         )
         .await;
+
+        let token_distribution_struct =
+            token_distribution(&owner1.token_distributor, fractional_nft_contract).await;
+        assert_eq!(
+            provider
+                .get_contract_asset_balance(
+                    &Bech32ContractId::from(token_distributor_contract),
+                    AssetId::from(*asset_contract)
+                )
+                .await
+                .unwrap(),
+            purchase_amount * token_price
+        );
+        assert!(matches!(
+            token_distribution_struct.clone().unwrap().state,
+            DistributionState::Distributed()
+        ));
+
         buyback(
-            purchase_amount * token_price,
+            purchase_amount * (token_price + 1),
             &owner1.token_distributor,
             asset_contract,
             fractional_nft_contract,
-            token_price,
+            token_price + 1,
         )
         .await;
 
+        let token_distribution_struct =
+            token_distribution(&owner1.token_distributor, fractional_nft_contract).await;
         assert_eq!(
-            wallet_balance(asset_contract, &owner2.wallet).await,
-            asset_supply - (purchase_amount * token_price)
+            provider
+                .get_contract_asset_balance(
+                    &Bech32ContractId::from(token_distributor_contract),
+                    AssetId::from(*asset_contract)
+                )
+                .await
+                .unwrap(),
+            (purchase_amount * token_price) + (purchase_amount * (token_price + 1))
         );
-        assert_eq!(
-            wallet_balance(fractional_nft_contract, &owner2.wallet).await,
-            purchase_amount
-        );
-
-        sell(
-            purchase_amount - 1,
-            &owner2.token_distributor,
-            fractional_nft_contract,
-        )
-        .await;
-
-        assert_eq!(
-            wallet_balance(asset_contract, &owner2.wallet).await,
-            asset_supply - token_price
-        );
-        assert_eq!(
-            wallet_balance(fractional_nft_contract, &owner2.wallet).await,
-            1
-        );
+        assert!(matches!(
+            token_distribution_struct.clone().unwrap().state,
+            DistributionState::Buyback()
+        ));
     }
 }
 
@@ -218,139 +220,23 @@ mod revert {
     async fn when_token_distribution_does_not_exist() {
         let (
             _deployer,
-            _owner1,
-            owner2,
+            owner1,
+            _owner2,
             _token_distributor_contract,
-            _fractional_nft_contract,
+            fractional_nft_contract,
             _nft_contract,
             asset_contract,
         ) = setup().await;
-        let (_reserve_price, _token_price, _token_supply, purchase_amount, asset_supply) =
+        let (_reserve_price, token_price, _token_supply, purchase_amount, asset_supply) =
             defaults().await;
 
-        mint_and_send_to_address(
-            asset_supply,
-            &owner2.asset,
-            Address::new(*owner2.wallet.address().hash()),
-        )
-        .await;
-
-        sell(purchase_amount, &owner2.token_distributor, asset_contract).await;
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "InvalidState")]
-    async fn when_not_accepting_returns() {
-        let (
-            _deployer,
-            owner1,
-            owner2,
-            _token_distributor_contract,
-            fractional_nft_contract,
-            nft_contract,
-            asset_contract,
-        ) = setup().await;
-        let (reserve_price, token_price, token_supply, purchase_amount, asset_supply) =
-            defaults().await;
-
-        let owner_identity = Identity::Address(owner1.wallet.address().into());
-        let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
-
-        mint(1, &owner1.nft, owner_identity.clone()).await;
-        approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
-        create(
-            &owner1.token_distributor,
-            asset_contract,
-            fractional_nft_contract,
-            nft_contract,
-            Some(reserve_price),
-            Some(owner_identity.clone()),
-            token_price,
-            token_supply,
-            0,
-        )
-        .await;
-        mint_and_send_to_address(
-            asset_supply,
-            &owner2.asset,
-            Address::new(*owner2.wallet.address().hash()),
-        )
-        .await;
         mint_and_send_to_address(
             asset_supply,
             &owner1.asset,
             Address::new(*owner1.wallet.address().hash()),
         )
         .await;
-        purchase(
-            purchase_amount,
-            &owner2.token_distributor,
-            asset_contract,
-            fractional_nft_contract,
-            token_price,
-        )
-        .await;
 
-        sell(
-            purchase_amount,
-            &owner2.token_distributor,
-            fractional_nft_contract,
-        )
-        .await;
-    }
-
-    #[tokio::test]
-    #[should_panic(expected = "InvalidAssetTransfer")]
-    async fn when_incorrect_asset_type() {
-        let (
-            _deployer,
-            owner1,
-            owner2,
-            _token_distributor_contract,
-            fractional_nft_contract,
-            nft_contract,
-            asset_contract,
-        ) = setup().await;
-        let (reserve_price, token_price, token_supply, purchase_amount, asset_supply) =
-            defaults().await;
-
-        let owner_identity = Identity::Address(owner1.wallet.address().into());
-        let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
-
-        mint(1, &owner1.nft, owner_identity.clone()).await;
-        approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
-        create(
-            &owner1.token_distributor,
-            asset_contract,
-            fractional_nft_contract,
-            nft_contract,
-            Some(reserve_price),
-            Some(owner_identity.clone()),
-            token_price,
-            token_supply,
-            0,
-        )
-        .await;
-        mint_and_send_to_address(
-            asset_supply,
-            &owner2.asset,
-            Address::new(*owner2.wallet.address().hash()),
-        )
-        .await;
-        mint_and_send_to_address(
-            asset_supply,
-            &owner1.asset,
-            Address::new(*owner1.wallet.address().hash()),
-        )
-        .await;
-        purchase(
-            purchase_amount,
-            &owner2.token_distributor,
-            asset_contract,
-            fractional_nft_contract,
-            token_price,
-        )
-        .await;
         buyback(
             purchase_amount * token_price,
             &owner1.token_distributor,
@@ -359,24 +245,186 @@ mod revert {
             token_price,
         )
         .await;
+    }
 
-        let tx_params = TxParameters::new(None, Some(1_000_000), None);
-        let call_params = CallParameters::new(
-            Some(asset_supply - (purchase_amount * token_price)),
-            Some(AssetId::from(*asset_contract)),
+    #[tokio::test]
+    #[should_panic(expected = "NotTokenAdmin")]
+    async fn when_no_owner() {
+        let (
+            _deployer,
+            owner1,
+            owner2,
+            _token_distributor_contract,
+            fractional_nft_contract,
+            nft_contract,
+            asset_contract,
+        ) = setup().await;
+        let (reserve_price, token_price, token_supply, purchase_amount, asset_supply) =
+            defaults().await;
+
+        let owner_identity = Identity::Address(owner1.wallet.address().into());
+        let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
+
+        mint(1, &owner1.nft, owner_identity.clone()).await;
+        approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
+        create(
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            nft_contract,
+            Some(reserve_price),
             None,
-        );
+            token_price,
+            token_supply,
+            0,
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner2.asset,
+            Address::new(*owner2.wallet.address().hash()),
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner1.asset,
+            Address::new(*owner1.wallet.address().hash()),
+        )
+        .await;
+        purchase(
+            purchase_amount,
+            &owner2.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            token_price,
+        )
+        .await;
 
-        owner2
-            .token_distributor
-            .methods()
-            .sell(fractional_nft_contract)
-            .tx_params(tx_params)
-            .call_params(call_params)
-            .append_variable_outputs(1)
-            .set_contract_ids(&[Bech32ContractId::from(fractional_nft_contract)])
-            .call()
-            .await
-            .unwrap();
+        buyback(
+            purchase_amount * token_price,
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            token_price,
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "InvalidState")]
+    async fn when_not_distributed() {
+        let (
+            _deployer,
+            owner1,
+            owner2,
+            _token_distributor_contract,
+            fractional_nft_contract,
+            nft_contract,
+            asset_contract,
+        ) = setup().await;
+        let (reserve_price, token_price, token_supply, purchase_amount, asset_supply) =
+            defaults().await;
+
+        let owner_identity = Identity::Address(owner1.wallet.address().into());
+        let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
+
+        mint(1, &owner1.nft, owner_identity.clone()).await;
+        approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
+        create(
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            nft_contract,
+            Some(reserve_price),
+            Some(owner_identity.clone()),
+            token_price,
+            token_supply,
+            0,
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner2.asset,
+            Address::new(*owner2.wallet.address().hash()),
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner1.asset,
+            Address::new(*owner1.wallet.address().hash()),
+        )
+        .await;
+
+        buyback(
+            purchase_amount * token_price,
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            token_price,
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    #[should_panic(expected = "InvalidAssetTransfer")]
+    async fn when_incorrect_amount_provided() {
+        let (
+            _deployer,
+            owner1,
+            owner2,
+            _token_distributor_contract,
+            fractional_nft_contract,
+            nft_contract,
+            asset_contract,
+        ) = setup().await;
+        let (reserve_price, token_price, token_supply, purchase_amount, asset_supply) =
+            defaults().await;
+
+        let owner_identity = Identity::Address(owner1.wallet.address().into());
+        let fractional_nft_identity = Identity::ContractId(fractional_nft_contract);
+
+        mint(1, &owner1.nft, owner_identity.clone()).await;
+        approve(Some(fractional_nft_identity.clone()), &owner1.nft, 0).await;
+        create(
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            nft_contract,
+            Some(reserve_price),
+            Some(owner_identity.clone()),
+            token_price,
+            token_supply,
+            0,
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner2.asset,
+            Address::new(*owner2.wallet.address().hash()),
+        )
+        .await;
+        mint_and_send_to_address(
+            asset_supply,
+            &owner1.asset,
+            Address::new(*owner1.wallet.address().hash()),
+        )
+        .await;
+        purchase(
+            purchase_amount,
+            &owner2.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            token_price,
+        )
+        .await;
+
+        buyback(
+            (purchase_amount - 1) * token_price,
+            &owner1.token_distributor,
+            asset_contract,
+            fractional_nft_contract,
+            token_price,
+        )
+        .await;
     }
 }
