+++ fractional-NFT/project/contracts/token-distributor-contract/tests/functions/mod.rs
@@ -1,10 +1,2 @@
-mod buyback;
-mod create;
-mod end;
-mod purchase;
-mod purchase_admin;
-mod sell;
-mod set_reserve;
-mod set_token_price;
-mod token_distribution;
-mod withdraw;
+mod core;
+mod info;
